+++ src/confirmationModal/story.tsx
@@ -2,8 +2,10 @@ import React, { Component } from 'react'
 
 import { storiesOf } from '@storybook/react'
 import { withKnobs, boolean, text } from '@storybook/addon-knobs'
+import { action } from '@storybook/addon-actions'
 
 import ConfirmationModal, { ConfirmationModalProps } from 'confirmationModal'
+import confirmationModalDoc from './specifications/confirmationModal.md'
 
 const stories = storiesOf('ConfirmationModal', module)
 stories.addDecorator(withKnobs)
@@ -19,6 +21,12 @@ class ConfirmationModalOpener extends Component<ConfirmationModalProps> {
 
   closeConfirmationModal = () => {
     this.setState({ confirmationModalOpen: false })
+    this.props.onClose()
+  }
+
+  confirmConfirmationModal = () => {
+    this.setState({ confirmationModalOpen: false })
+    this.props.onConfirm()
   }
 
   render() {
@@ -28,6 +36,7 @@ class ConfirmationModalOpener extends Component<ConfirmationModalProps> {
         <ConfirmationModal
           {...this.props}
           onClose={this.closeConfirmationModal}
+          onConfirm={this.confirmConfirmationModal}
           isOpen={this.state.confirmationModalOpen}
         >
           <div>
@@ -40,26 +49,36 @@ class ConfirmationModalOpener extends Component<ConfirmationModalProps> {
   }
 }
 
-stories.add('Warning', () => (
-  <ConfirmationModalOpener
-    status={ConfirmationModal.STATUS.WARNING}
-    isOpen={false}
-    onClose={() => {}}
-    onConfirm={() => console.log('fhtagn')}
-    confirmLabel={text('confirmLabel', 'Fhtagn')}
-    large={boolean('large', false)}
-    closeOnEsc={boolean('closeOnEsc', true)}
-  />
-))
+stories.add(
+  'Warning',
+  () => (
+    <ConfirmationModalOpener
+      status={ConfirmationModal.STATUS.WARNING}
+      onClose={action('onClose')}
+      onConfirm={action('onConfirm')}
+      confirmLabel={text('confirmLabel', 'Fhtagn')}
+      large={boolean('large', false)}
+      closeOnEsc={boolean('closeOnEsc', true)}
+    />
+  ),
+  {
+    readme: { content: confirmationModalDoc },
+  },
+)
 
-stories.add('Reminder', () => (
-  <ConfirmationModalOpener
-    status={ConfirmationModal.STATUS.REMINDER}
-    isOpen={false}
-    onClose={() => {}}
-    onConfirm={() => console.log('fhtagn')}
-    confirmLabel={text('confirmLabel', 'Fhtagn')}
-    large={boolean('large', false)}
-    closeOnEsc={boolean('closeOnEsc', true)}
-  />
-))
+stories.add(
+  'Reminder',
+  () => (
+    <ConfirmationModalOpener
+      status={ConfirmationModal.STATUS.REMINDER}
+      onClose={action('onClose')}
+      confirmLabel={text('confirmLabel', 'Fhtagn')}
+      onConfirm={action('onConfirm')}
+      large={boolean('large', false)}
+      closeOnEsc={boolean('closeOnEsc', true)}
+    />
+  ),
+  {
+    readme: { content: confirmationModalDoc },
+  },
+)