+++ src/confirmationModal/style.ts
@@ -4,23 +4,26 @@ import { color, space, font } from '_utils/branding'
 const footerHeight = '96px' /* = padding + content */
 
 export default css`
-  .kirk-confirmationModal {
-    z-index: 999;
-    position: fixed;
-    top: 0;
-    right: 0;
-    width: 100%;
-    height: 100%;
-    overflow-y: auto;
+  :global(.kirk-modal.kirk-confirmationModal) {
     background-color: rgba(5, 71, 82, 0.95);
     color: ${color.white};
   }
 
-  .kirk-confirmationModal--large {
+  :global(.kirk-modal.kirk-confirmationModal .kirk-modal-dialog) {
+    background: transparent;
+    width: auto !important;
+    padding: 0;
+    max-width: initial;
+    margin: 0;
+    display: flex;
+    height: 100%;
+  }
+
+  :global(.kirk-modal.kirk-confirmationModal--large) {
     text-align: center;
   }
 
-  .kirk-confirmationModal-dialog {
+  :global(.kirk-modal .kirk-confirmationModal-dialog) {
     position: relative;
     margin: 0 auto;
     box-shadow: none;
@@ -30,22 +33,22 @@ export default css`
     padding-bottom: ${footerHeight};
   }
 
-  .kirk-confirmationModal--large .kirk-confirmationModal-dialog {
+  :global(.kirk-modal.kirk-confirmationModal--large .kirk-confirmationModal-dialog) {
     max-width: 600px;
   }
 
-  .kirk-confirmationModal-body {
+  :global(.kirk-modal .kirk-confirmationModal-body) {
     padding: ${space.xl};
     font-size: ${font.xl.size};
     line-height: ${font.xl.lineHeight};
   }
 
-  :global(.kirk-confirmationModal-icon) {
+  :global(.kirk-modal .kirk-confirmationModal-icon) {
     margin-top: ${space.xl};
     margin-left: ${space.xl};
   }
 
-  .kirk-confirmationModal-footer {
+  :global(.kirk-modal .kirk-confirmationModal-footer) {
     position: absolute;
     bottom: 0;
     left: 0;
@@ -57,26 +60,26 @@ export default css`
     background-color: ${color.warningBackground};
   }
 
-  .kirk-confirmationModal--large .kirk-confirmationModal-footer {
+  :global(.kirk-modal.kirk-confirmationModal--large .kirk-confirmationModal-footer) {
     position: relative;
     display: block;
     margin: 0 auto;
     background: none;
   }
 
-  .kirk-confirmationModal-footer {
+  :global(.kirk-modal .kirk-confirmationModal-footer) {
     justify-content: flex-end;
   }
 
-  .kirk-confirmationModal--hasCloseButton .kirk-confirmationModal-footer {
+  :global(.kirk-modal.kirk-confirmationModal--hasCloseButton .kirk-confirmationModal-footer) {
     justify-content: space-between;
   }
 
-  :global(.kirk-confirmationModal--large .kirk-confirmationModal-footer .kirk-button) {
+  :global(.kirk-modal.kirk-confirmationModal--large .kirk-confirmationModal-footer .kirk-button) {
     margin: ${space.l} ${space.xl} 0;
   }
 
-  :global(.kirk-confirmationModal-closeButton) {
+  :global(.kirk-modal .kirk-confirmationModal-closeButton) {
     vertical-align: bottom; /* vertical alignment to be fixed in Button */
     border: 0; /* content alignment to be fixed in Button */
   }