+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- [...]
+- **[BREAKING CHANGE]** `ConfirmationModal` & `SuccessModal` use `Modal` (`onConfirm` on `SuccessModal` is replaced with `onClose`)
 
 # v2.0.2 (12/04/2019)